Adding `InGroup` associatedtype constraint on `GroupElement`, each instance of a groups element now has to reference an instance of its group type.

+++ Source/Code/Group/GroupTypes/FiniteGroup.swift
@@ -9,5 +9,5 @@
 import Foundation
 
 public protocol FiniteGroup: Group {
-    var order: Int64 { get }
+    var order: Int { get }
 }
